services/pdf-image/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyMuPDF==1.23.0
python-docx==0.8.11
reportlab==4.0.4
Pillow>=10.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
pdfservices-sdk==4.2.0
pdf2image==1.17.0
svglib==1.5.1
cairosvg==2.7.1
wand==0.6.13
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
